+++ CHANGELOG.md
@@ -6,6 +6,10 @@
 - Add `correlationId` to text, custom and cancel response packets
 - Allow to use async `onDisconnect` and `onError`
 
+### Fixed
+
+- Generate proto packet just once to dispatching it to the bidirectional stream
+
 ## [1.7.0] - 2023-09-01
 
 ### Added